particles works now except at very certain zoom ins it turns 2d (when fire not at origin). Still working on water

+++ KPTB_167/ParticleSpawn.h
@@ -25,7 +25,7 @@ class ParticleSpawn
 public:
 	struct Particle {
 		glm::vec3 pos, speed;
-		GLubyte r, g, b, a; // Color
+		float r, g, b, a; // Color
 		float life; // Remaining life of the particle. if < 0 : dead and unused.
 
 		Particle()
